+++ packages/safe-apps-web3modal/src/modal.ts
@@ -43,6 +43,11 @@ export class SafeAppWeb3Modal extends Web3Modal {
   };
 
   public async isSafeApp(): Promise<boolean> {
+    // check if we're in an iframe
+    if (window?.parent === window) {
+      return false;
+    }
+
     const safe = await this.getConnectedSafe();
 
     return !!safe;