+++ packages/safe-apps-web3-react/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @gnosis.pm/safe-apps-web3-react
 
+## 0.6.4
+
+### Patch Changes
+
+- 2b56d7e: move unecessary dependencies to dev-dependencies
+- 8137ae6: dependency bump
+  in isSafeApp check, check if running inside an iframe
+
+  Sent from my iPhone
+
+- Updated dependencies [2b56d7e]
+- Updated dependencies [9a4d3a4]
+  - @gnosis.pm/safe-apps-provider@0.8.1
+  - @gnosis.pm/safe-apps-sdk@5.0.0
+
+## 0.6.4-next.2
+
+### Patch Changes
+
+- Updated dependencies [9a4d3a4]
+  - @gnosis.pm/safe-apps-sdk@5.0.0-next.1
+  - @gnosis.pm/safe-apps-provider@0.8.1-next.1
+
+## 0.6.4-next.1
+
+### Patch Changes
+
+- 2b56d7e: move unecessary dependencies to dev-dependencies
+- Updated dependencies [2b56d7e]
+  - @gnosis.pm/safe-apps-provider@0.8.1-next.0
+  - @gnosis.pm/safe-apps-sdk@4.3.1-next.0
+
+## 0.6.4-next.0
+
+### Patch Changes
+
+- 8137ae6: dependency bump
+  in isSafeApp check, check if running inside an iframe
+
+  Sent from my iPhone
+
 ## 0.6.3
 
 ### Patch Changes